requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
cffi==1.10.0
cryptography==2.0.3
enum34==1.1.6
Fabric==1.13.2
idna==2.6
ipaddress==1.0.18
Jinja2==2.9.6
MarkupSafe==1.0
paramiko==2.2.1
pyasn1==0.3.2
pycparser==2.18
PyNaCl==1.1.2
PyYAML==3.12
six==1.10.0
